AP-9465: add human readable service name
Add a human readable service name to the connection informations that
can be used in error messages.

Note: S3/Azure use the same cloud connection informations class instead of
implementing a service specific version... This is why we can't use a method
override in the implementation and have to use a ugly workaround.

AP-9465 (Google Cloud Storage Support)